+++ src/main/java/de/codescape/jira/plugins/scrumpoker/action/ScrumPokerGettingStartedAction.java
@@ -15,7 +15,7 @@ public class ScrumPokerGettingStartedAction extends AbstractScrumPokerAction imp
     }
 
     @Override
-    protected String doExecuteInternal() {
+    protected String doExecute() {
         return SUCCESS;
     }