+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * chore: annual review for Data Center approval
 * feat: introduce XSRF protection for health check
 * feat: introduce XSRF protection for error log
+* feat: introduce XSRF protection for global configuration
 
 ### [21.04.0] - 2021-04-19