+++ docs/changelog.md
@@ -15,6 +15,7 @@ category: Administration
 * feat: introduce XSRF protection for health check
 * feat: introduce XSRF protection for error log
 * feat: introduce XSRF protection for global configuration
+* feat: introduce XSRF protection for project configuration
 
 ### [21.04.0] - 2021-04-19