+++ docs/changelog.md
@@ -11,6 +11,12 @@ category: Administration
 
 ### [Unreleased]
 
+* ...
+
+### [21.05.0] - 2021-05-09
+
+**Improve XSRF security for all configuration pages**
+
 * chore: annual review for Data Center approval
 * feat: introduce XSRF protection for health check
 * feat: introduce XSRF protection for error log
@@ -66,7 +72,8 @@ Older versions have been moved into separate changelog documents grouped by thei
 * [Scrum Poker for Jira 2.x](/changelog-2x)
 * [Scrum Poker for Jira 1.x](/changelog-1x)
 
-[Unreleased]: https://github.com/codescape/jira-scrum-poker/compare/21.04.0...HEAD
+[Unreleased]: https://github.com/codescape/jira-scrum-poker/compare/21.05.0...HEAD
+[21.05.0]: https://github.com/codescape/jira-scrum-poker/compare/21.04.0...21.05.0
 [21.04.0]: https://github.com/codescape/jira-scrum-poker/compare/21.02.0...21.04.0
 [21.02.0]: https://github.com/codescape/jira-scrum-poker/compare/21.01.1...21.02.0
 [21.01.1]: https://github.com/codescape/jira-scrum-poker/compare/21.01.0...21.01.1